Fixed backend logic for adding and removing a user to/from a group

+++ backend/Lookout/src/main/kotlin/com/lookout/Lookout/services/GroupService.kt
@@ -36,6 +36,11 @@ class GroupService(private val groupRepository: GroupRepository, private val use
     fun addMember(groupId: Long, userId: Long) {
         val group = findById(groupId) ?: throw IllegalArgumentException("Group not found with id: $groupId")
         val user = userService.findById(userId).orElseThrow { IllegalArgumentException("User not found with id: $userId") }
+
+        if (groupRepository.countMembersInGroup(group.id, user.id) > 0) {
+            throw IllegalArgumentException("User is already in the group")
+        }
+
         groupRepository.addMemberToGroup(group.id, user.id)
     }
 
@@ -43,6 +48,11 @@ class GroupService(private val groupRepository: GroupRepository, private val use
     fun removeMember(groupId: Long, userId: Long) {
         val group = findById(groupId) ?: throw IllegalArgumentException("Group not found with id: $groupId")
         val user = userService.findById(userId).orElseThrow { IllegalArgumentException("User not found with id: $userId") }
+
+        if (groupRepository.countMembersInGroup(group.id, user.id).toInt() == 0) {
+            throw IllegalArgumentException("User is not in the group")
+        }
+
         groupRepository.removeMemberFromGroup(group.id, user.id)
     }
